@@ -2,8 +2,8 @@ import Link from "next/link";
 
 export const ButtonLink = ({ destination, label }) => {
   return (
-    <Link href={destination}>
-      <a className="btn">{label}</a>
+    <Link href={destination} className="btn">
+      {label}
     </Link>
   );
 };
