@@ -20,48 +20,50 @@ export const PropertyCard = ({
   petFriendly,
 }) => {
   return (
-    <Link href={destination}>
-      <a className="border-2 border-slate-300 p-5 block bg-slate-100 hover:bg-slate-200">
-        <div className="flex w-full">
-          <Image
-            src={image}
-            height="200px"
-            width="300px"
-            objectFit="cover"
-            alt=""
-          />
+    <Link
+      href={destination}
+      className="border-2 border-slate-300 p-5 block bg-slate-100 hover:bg-slate-200"
+    >
+      <div className="flex w-full relative h-[200px]">
+        <Image
+          src={image}
+          priority="low"
+          fill
+          alt=""
+          sizes="300px"
+          style={{ objectFit: "cover" }}
+        />
+      </div>
+      <div className="mt-3 text-lg font-bold">{title}</div>
+      <div className="text-lg">£{numeral(price).format("0,0")}</div>
+      <div className="flex justify-between text-sm mt-3">
+        <div>
+          <FontAwesomeIcon icon={faBathtub} />
+          <span className="pl-2">{bathrooms} bathrooms</span>
         </div>
-        <div className="mt-3 text-lg font-bold">{title}</div>
-        <div className="text-lg">£{numeral(price).format("0,0")}</div>
+        <div>
+          <FontAwesomeIcon icon={faBed} />
+          <span className="pl-2">{bedrooms} bedrooms</span>
+        </div>
+      </div>
+      {(!!hasParking || !!petFriendly) && (
         <div className="flex justify-between text-sm mt-3">
           <div>
-            <FontAwesomeIcon icon={faBathtub} />
-            <span className="pl-2">{bathrooms} bathrooms</span>
+            {!!hasParking && (
+              <>
+                <FontAwesomeIcon icon={faCar} /> parking available
+              </>
+            )}
           </div>
           <div>
-            <FontAwesomeIcon icon={faBed} />
-            <span className="pl-2">{bedrooms} bedrooms</span>
+            {!!petFriendly && (
+              <>
+                <FontAwesomeIcon icon={faDog} /> pet friendly
+              </>
+            )}
           </div>
         </div>
-        {(!!hasParking || !!petFriendly) && (
-          <div className="flex justify-between text-sm mt-3">
-            <div>
-              {!!hasParking && (
-                <>
-                  <FontAwesomeIcon icon={faCar} /> parking available
-                </>
-              )}
-            </div>
-            <div>
-              {!!petFriendly && (
-                <>
-                  <FontAwesomeIcon icon={faDog} /> pet friendly
-                </>
-              )}
-            </div>
-          </div>
-        )}
-      </a>
+      )}
     </Link>
   );
 };
